+++ src/runtime/server/ssr/endRequest.js
@@ -1,0 +1,9 @@
+export default ({ ctx, html }) =>
+  ctx.succeed({
+    statusCode: 200,
+    headers: {
+      "Content-Type": "text/html",
+      "Access-Control-Allow-Origin": "*",
+    },
+    body: html,
+  });
