+++ src/runtime/server/ssr/index.js
@@ -1,0 +1,16 @@
+import { compose, ifElse } from "ramda";
+
+import endRequest from "./endRequest";
+import redirect from "./redirect";
+import renderBody from "./renderBody";
+import withApp from "./withApp";
+import withHtml from "./html";
+import withMeta from "./withMeta";
+
+export const shouldRedirect = ifElse(
+  ({ context }) => context.url,
+  redirect,
+  compose(endRequest, withHtml, withMeta),
+);
+
+export default App => compose(shouldRedirect, renderBody, withApp(App));
