+++ src/runtime/server/ssr/withApp.js
@@ -1,0 +1,7 @@
+export default App =>
+  (event, ctx) => ({
+    path: event.path,
+    ctx,
+    context: {},
+    App,
+  });
