+++ src/runtime/server/ssr/withMeta.js
@@ -1,0 +1,6 @@
+import Helmet from "react-helmet";
+
+export default x => ({
+  ...x,
+  metadata: Helmet.renderStatic(),
+});
